+++ src/components/Checkbox/composables/index.ts
@@ -48,9 +48,6 @@ export function useCheckboxClasses(props: PropTypes) {
       display: 'inline-flex',
       alignItems: 'items-center',
       justifyContent: 'justify-center',
-      // position: "relative",
-      // zIndex: "z-10",
-      // 16 - 20 - 24 - 28 - 32
       width: {
         'w-4': props.size.value === 'xs',
         'w-5': props.size.value === 'sm',
@@ -98,25 +95,6 @@ export function useCheckboxClasses(props: PropTypes) {
     }
     return Object.values({ ...classes })
   })
-
-  // const descriptionClasses = computed(() => {
-  //   const classes: TwClassInterface = {
-  //     fontSize: "text-sm",
-  //     lineHeight: "leading-normal",
-  //     margin: "mt-1",
-  //     textColor:
-  //       attrs.disabled !== undefined && attrs.disabled.value
-  //         ? "text-secondary-200 dark:text-secondary-700"
-  //         : classNames({
-  //           "text-secondary-600 dark:text-secondary-400":
-  //             props.state.value === null,
-  //           "text-warning-500": props.state.value === "warning",
-  //           "text-danger-500": props.state.value === false,
-  //           "text-success-500": props.state.value === true,
-  //         }),
-  //   };
-  //   return classNames(Object.values({ ...classes }));;
-  // });
 
   return { inputClasses, spanClasses, iconClasses }
 }