+++ src/components/Input/input-types.ts
@@ -1,4 +1,4 @@
 export type Clear = false | true
-export type ModelValue = string | number
+export type ModelValue = string | number | undefined
 export type Description = string | null
 export type IconSize = 'xs' | 'sm' | 'md' | 'lg' | 'xl'