+++ src/components/Select/LuiSelect.stories.ts
@@ -89,163 +89,63 @@ export const Default: Story = {
         testChange
       }
     },
-    template: defaultTemplate
+    template: `
+    <div class="flex space-x-10 h-[1400px]">
+      <div>
+        <lui-select v-bind="args" placeholder="TEST!" @change="testChange" :options="['a','b','c']" />
+      </div>
+      <div>
+        <lui-select>
+          <template #prepend>
+            <i class="ri-home-line"/>
+          </template>
+          <lui-option v-for="i in 5" :key="i" :text="i + 'okeygoogle ve bazi baska kelimeler sen bilir misin bilmem '" :value="i" :disabled="i == 1 || i == 2">
+            <template #append>
+              <i class="ri-home-line"/>
+            </template>
+          </lui-option>
+        </lui-select>
+      </div>
+      <div class="mt-[600px]">
+        <lui-select v-bind="args">
+          <lui-option text="Second option first item" value="6"></lui-option>
+          <lui-option text="Second option 2. item" value="7"></lui-option>
+          <lui-option text="Second option 3.. item" value="11"></lui-option>
+        </lui-select>
+      </div>
+    </div>
+  `
   })
 }
 
-// export const WithObjectArrays: Story = {
-//   render: (args) => ({
-//     components: { LuiSelect },
-//     setup() {
-//       const names = [
-//         { text: 'bayhan', value: '9' },
-//         { text: 'sercan', value: '10' },
-//         { text: 'serkan', value: '12' },
-//         { text: 'rahmi', value: '4' },
-//         { text: 'can', value: '7' }
-//       ]
-//       const selectedName = ref(names[0].text)
-//       return {
-//         args,
-//         names,
-//         selectedName
-//       }
-//     },
-//     template: `
-//     <div class="p-2">
-//       SelectedName: {{selectedName}}
-//     </div>
-//     <lui-select :options="names" v-model="selectedName"/>
-//   `
-//   })
-// }
-// export const WithSlots: Story = {
-//   render: (args) => ({
-//     components: { LuiSelect, LuiOption },
-//     setup() {
-//       const items = ['javascript', 'html', 'css', 'vuejs']
-//       const selectedItem = ref(items[1])
-//       return { args, items, selectedItem }
-//     },
-//     template: `
-//     <div class="w-[800px]">
-//       <lui-select v-model="selectedItem">
-//         <lui-option v-for="item in items" :key="item" :id="item" :text="item" />
-//       </lui-select>
-//     </div>
-//   `
-//   })
-// }
-// export const MaxHeight: Story = {
-//   render: (args) => ({
-//     components: { LuiSelect, LuiOption },
-//     setup() {
-//       const cities = [
-//         'Adana',
-//         'Adiyaman',
-//         'Afyon',
-//         'Agri',
-//         'Aksaray',
-//         'Amasya',
-//         'Ankara',
-//         'Antalya',
-//         'Ardahan',
-//         'Artvin',
-//         'Aydin',
-//         'Balikesir',
-//         'Bartin',
-//         'Batman',
-//         'Bayburt',
-//         'Bilecik',
-//         'Bingol',
-//         'Bitlis',
-//         'Bolu',
-//         'Burdur',
-//         'Bursa',
-//         'Canakkale',
-//         'Cankiri',
-//         'Corum',
-//         'Denizli',
-//         'Diyarbakir',
-//         'Duzce',
-//         'Edirne',
-//         'Elazig',
-//         'Erzincan',
-//         'Erzurum',
-//         'Eskisehir',
-//         'Gaziantep',
-//         'Giresun',
-//         'Gumushane',
-//         'Hakkari',
-//         'Hatay',
-//         'Igdir',
-//         'Isparta',
-//         'Istanbul',
-//         'Izmir',
-//         'Kahramanmaras',
-//         'Karabuk',
-//         'Karaman',
-//         'Kars',
-//         'Kastamonu',
-//         'Kayseri',
-//         'Kilis',
-//         'Kirikkale',
-//         'Kirklareli',
-//         'Kirsehir',
-//         'Kocaeli',
-//         'Konya',
-//         'Kutahya',
-//         'Malatya',
-//         'Manisa',
-//         'Mardin',
-//         'Mersin',
-//         'Mugla',
-//         'Mus',
-//         'Nevsehir',
-//         'Nigde',
-//         'Ordu',
-//         'Osmaniye',
-//         'Rize',
-//         'Sakarya',
-//         'Samsun',
-//         'Sanliurfa',
-//         'Siirt',
-//         'Sinop',
-//         'Sirnak',
-//         'Sivas',
-//         'Tekirdag',
-//         'Tokat',
-//         'Trabzon',
-//         'Tunceli',
-//         'Usak',
-//         'Van',
-//         'Yalova',
-//         'Yozgat',
-//         'Zonguldak'
-//       ]
-//       const selectedCity = ref(cities[0])
-//       return { args, cities, selectedCity }
-//     },
-//     template: `
-//       <lui-select v-model="selectedCity" :options="cities" size="xl" />
-//   `
-//   })
-// }
-
-const prependTemplate = `
-<lui-select>
-<template #prepend>
-  <i class="ri-home-line"/>
-</template>
-<lui-option v-for="i in 5" :key="i" :text="i + 'okeygoogle'" :value="i" :disabled="i == 1 || i == 2">
-  <template #append>
-    <i class="ri-home-line"/>
-  </template>
-</lui-option>
-</lui-select>
-`
-export const Prepend: Story = {
-  render: () => ({
+export const WithObjectArrays: Story = {
+  render: (args) => ({
+    components: { LuiSelect },
+    setup() {
+      const names = [
+        { text: 'bahyan', value: '9' },
+        { text: 'sercan', value: '10' },
+        { text: 'serkan', value: '12' },
+        { text: 'rahmi', value: '4' },
+        { text: 'can', value: '7' }
+      ]
+      const selectedName = ref(names[0].text)
+      return {
+        args,
+        names,
+        selectedName
+      }
+    },
+    template: `
+    <div class="p-2">
+      SelectedName: {{selectedName}}
+    </div>
+    <lui-select :options="names" v-model="selectedName"/>
+  `
+  })
+}
+export const WithSlots: Story = {
+  render: (args) => ({
     components: { LuiSelect, LuiOption },
     template: prependTemplate
   }),